migrations/mainApp/1_20240617232051_add_type_field.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "profilerecords" ADD "type" VARCHAR(10) NOT NULL  DEFAULT 'telnet';"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "profilerecords" DROP COLUMN "type";"""
