requirements.txt
requests
tortoise-orm
sanic
sanic-cors
aerich